marvel_vector_db/requirements.txt
# Marvel Vector Database Requirements

# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
langchain-ollama>=0.1.0

# Vector database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Document processing
unstructured[all-docs]>=0.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Image processing
Pillow>=10.0.0

# Audio processing
yt-dlp>=2023.7.0
whisper>=1.1.0
soundfile>=0.12.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

